backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
transformers>=4.30.0
torch>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.5
SQLAlchemy>=1.4.0
aiosqlite>=0.17.0